+++ src/main/java/se/mau/myhappyplants/library/UserPlantRepository.java
@@ -14,31 +14,31 @@ import java.util.Optional;
  * Provides queries to list and manage a user's plant collection.
  */
 @Repository
-public interface UserPlantRepository extends JpaRepository<UserPlant, Long> {
+public interface UserPlantRepository extends JpaRepository<UserPlant, Integer> {
     /**
      * Hitta alla växter som tillhör en specifik användare
      */
-    List<UserPlant> findByUserId(Long userId);
+    List<UserPlant> findByUserId(int userId);
 
     /**
      * Hitta en specifik växt för en användare
      */
-    Optional<UserPlant> findByIdAndUserId(Long id, Long userId);
+    Optional<UserPlant> findByIdAndUserId(int id, int userId);
 
     /**
      * Hitta växter baserat på namn (för en specifik användare)
      */
-    List<UserPlant> findByUserIdAndPlantNameContainingIgnoreCase(Long userId, String plantName);
+    List<UserPlant> findByUserIdAndPlantNameContainingIgnoreCase(int userId, String plantName);
 
     /**
      * Hitta växter med en specifik tagg
      */
-    List<UserPlant> findByUserIdAndTagId(Long userId, Long tagId);
+    List<UserPlant> findByUserIdAndTagId(int userId, int tagId);
 
     /**
      * Räkna antal växter för en användare
      */
-    long countByUserId(Long userId);
+    long countByUserId(int userId);
 
     /**
      * Hitta växter baserat på Perenual API ID
@@ -48,10 +48,10 @@ public interface UserPlantRepository extends JpaRepository<UserPlant, Long> {
     /**
      * Hämta växter sorterade alfabetiskt (A-Z)
      */
-    List<UserPlant> findByUserIdOrderByPlantNameAsc(Long userId);
+    List<UserPlant> findByUserIdOrderByPlantNameAsc(int userId);
 
     /**
      * Hämta växter sorterade omvänt alfabetiskt (Z-A)
      */
-    List<UserPlant> findByUserIdOrderByPlantNameDesc(Long userId);
+    List<UserPlant> findByUserIdOrderByPlantNameDesc(int userId);
 }
